% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDocumentTermMatrix.R
\name{asDocumentTermMatrix}
\alias{asDocumentTermMatrix}
\title{Document-Term Matrix}
\usage{
asDocumentTermMatrix(input, vect.vocab = NULL,
  stopwords = character(0), stemming = NULL, type = c("dgCMatrix",
  "dgTMatrix", "lda_c"))
}
\arguments{
\item{input}{a character vector.}

\item{vect.vocab}{a vocabulary created with \code{\link[text2vec:vectorizers]{vocab_vectorizer}}. If \code{NULL}, the vocabulary is created from the input. See example for a typical use case.}

\item{stopwords}{character vector of stopwords to exclude when creating the vocabulary. \code{tm::stopwords("de")} provides German stopwords.}

\item{stemming}{\code{NULL} for no stemming and \code{"de"} for stemming using the German porter stemmer.}

\item{type}{character, one of c("dgCMatrix", "dgTMatrix", "lda_c") taken from \code{\link[text2vec]{create_dtm}}. \code{dgCMatrix} are useful for glmnet; \code{dgTMatrix} matrix refers to sparse matrices in triplet form, i.e. positions of all non-zero values are stored (easier to work with, but non-unique).}
}
\value{
A list with two elements
\describe{
  \item{dtm}{a sparse document-term-matrix, depending on the \code{type}-parameter}
  \item{vect.vocab}{a vocabulary that can be inserted as \code{vect.vocab} to build a document term matrix on new data with the same vocabulary.}
}
}
\description{
Constructs a document-term matrix.
}
\examples{
x <- c("Verkauf von Schreibwaren", "Verkauf", "Schreibwaren")
asDocumentTermMatrix(x)
asDocumentTermMatrix(x, type = "dgTMatrix")
asDocumentTermMatrix(x, stopwords = tm::stopwords("de"))

(x <- c("Verkauf von B\\xfcchern, Schreibwaren", "Fach\\xe4rzin f\\xfcr Kinder- und Jugendmedizin im \\xf6ffentlichen Gesundheitswesen", "Industriemechaniker", "Dipl.-Ing. - Agrarwirtschaft (Landwirtschaft)"))
x <- stringPreprocessing(x)
dtm <- asDocumentTermMatrix(x, stemming = "de")
print(dtm$dtm)
dimnames(dtm$dtm)[[2]]

# use the newly created vocab_vectorizer
(x <- stringPreprocessing(c("WILL NOT SHOW UP", "Verkauf von B\\xfcchern, Schreibwaren", "Fach\\xe4rzin f\\xfcr Kinder- und Jugendmedizin")))
asDocumentTermMatrix(x, vect.vocab = dtm$vect.vocab, stopwords = character(0), stemming = "de")$dtm
}
\seealso{
\url{http://text2vec.org/vectorization.html} for details on the implementation used here,
another implementation \code{\link[tm:TermDocumentMatrix]{TermDocumentMatrix}} is slower
}
