% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{coding_index_excerpt}
\alias{coding_index_excerpt}
\title{An excerpt from the Gesamtberufsliste der BA}
\format{
A data frame with 90 rows and 5 variables:
\describe{
  \item{Berufsbenennungen}{neutral short title}
  \item{bezMale}{male long title}
  \item{bezFemale}{female long title}
  \item{Code}{Code from the German Classification of Occupations (KldB 2010)}
  \item{count_categories}{An indicator of job title ambiguity. Only used within function \code{\link{predictWithCodingIndex}}}
}
}
\source{
Bundesagentur für Arbeit (2019). Gesamtberufsliste der Bundesagentur für Arbeit. Stand: 03.01.2019. The \code{Gesamtberufsliste der BA} is available at \url{https://download-portal.arbeitsagentur.de/files/}.

The function \code{\link{prepare_German_coding_index_Gesamtberufsliste_der_BA}} was used to process the downloaded \code{.xlsx}-file and prepare this coding index. The resulting coding index has 27853 rows. We selected 90 rows that are related to the \code{\link{occupations}} dataset.
}
\usage{
coding_index_excerpt
}
\description{
90 selected job titles from the \code{Gesamtberufsliste der BA}. This dataset is not complete and only used to demonstrate the functionality of the package. We recommend that users build their own coding index. \code{\link{prepare_German_coding_index_Gesamtberufsliste_der_BA}} can be used to build a German coding index.
}
\keyword{datasets}
