% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSimilarityTableSubstring.R
\name{createSimilarityTableSubstring}
\alias{createSimilarityTableSubstring}
\title{Similarity Table with Coding index}
\usage{
createSimilarityTableSubstring(unique.string, coding_index_w_codes,
  coding_index_without_codes)
}
\arguments{
\item{unique.string}{a character vector (usually unique(answer))}

\item{coding_index_w_codes}{a data.table with columns "title" and "Code".}

\item{coding_index_without_codes}{a character vector of additional titles}
}
\value{
a list with elements
\describe{
  \item{dist_table_w_code}{a data.table with colummns \code{intString}, \code{dictString.title}, \code{dictString.Code}, \code{dist}}
  \item{dist_table_without_code}{\code{NULL} or a data.table with colummns \code{intString}, \code{dictString}, \code{dist}}
  \item{vect_vocab}{see \code{link{asDocumentTermMatrix}}}
}
}
\description{
Calculate SUBSTRING similarity between \code{unique.string} and \code{(coding_index_w_codes, coding_index_without_codes)}. unique.string and coding_index are similar if coding_index is a substring of unique.string.
}
\details{
Special function for similarity-based reasoning: creates distance data with substring-method
}
\examples{
## Prepare coding index
# write female titles beneath the male title
coding_index <- rbind(coding_index_excerpt[, list(title = bezMale, Code)],
                      coding_index_excerpt[, list(title = bezFemale, Code)])
# standardize titles from the coding index
coding_index <- coding_index[,title := stringPreprocessing(title)]
# drop duplicate lines, might be suboptimal because we keep each title and its associated code only a single time. This means we delete duplicates and the associated, possibly relevant codes.
coding_index <- coding_index[!duplicated(title)]

(x <- c("Abgeordneter", "Abgeordneter", "Abgeordnete", "abgeordnet", "FSJ", "FSJ2", "Industriemechaniker", "Dipl.-Ing. - Agrarwirtschaft (Landwirtschaft)"))
createSimilarityTableSubstring(unique.string = stringPreprocessing(x),
                               coding_index_w_codes = coding_index,
                               coding_index_without_codes = frequent_phrases)
}
\seealso{
\code{\link{trainSimilarityBasedReasoning}}, \code{\link{createSimilarityTableWordwiseStringdist}}, \code{\link{createSimilarityTableStringdist}}, \code{\link{createSimilarityTableStringdist}}
}
