% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSimilarityTableWordwiseStringdist.R
\name{createSimilarityTableWordwiseStringdist}
\alias{createSimilarityTableWordwiseStringdist}
\title{Wordwise Similarity Table with Coding index}
\usage{
createSimilarityTableWordwiseStringdist(
  unique.string,
  coding_index_w_codes,
  coding_index_without_codes,
  preprocessing,
  dist.control = list(method = "osa", weight = c(d = 1, i = 1, s = 1, t = 1)),
  threshold = 1
)
}
\arguments{
\item{unique.string}{a character vector (usually unique(answer))}

\item{coding_index_w_codes}{a data.table with columns "title" and "Code".}

\item{coding_index_without_codes}{a character vector of additional titles}

\item{preprocessing}{a list with elements
\describe{
  \item{stopwords}{a character vector, use \code{tm::stopwords("de")} for German stopwords.}
  \item{stemming}{\code{NULL} for no stemming and \code{"de"} for stemming using the German porter stemmer.}
  \item{strPreprocessing}{\code{TRUE} if \code{\link{stringPreprocessing}} shall be used.}
  \item{removePunct}{\code{TRUE} if \code{\link[tm]{removePunctuation}} shall be used.}
}}

\item{dist.control}{a list that will be passed to \code{\link[stringdist:stringdist]{stringdistmatrix}}. Currently only two elements are implemented:
\describe{
  \item{method}{Method for distance calculation.}
  \item{weight}{For method='osa' or 'dl'.}
}}

\item{threshold}{All entries with distance above this threshold will be removed from the result}
}
\value{
a list with elements
\describe{
  \item{dist_table_w_code}{a data.table with colummns \code{intString}, \code{dictString.title}, \code{dictString.Code}, \code{dist}}
  \item{dist_table_without_code}{\code{NULL} or a data.table with colummns \code{intString}, \code{dictString}, \code{dist}}
  \item{vect_vocab}{see \code{link{asDocumentTermMatrix}}}
}
}
\description{
Calculate string similarity on a word basis between \code{unique.string} and \code{(coding_index_w_codes, coding_index_without_codes)}.
}
\details{
Special function for similarity-based reasoning: creates WORDWISE!!!!! distance data with osa-method c(d = 1, i = 1, s = 1, t = 1)
--> allows to correct 1 letter in a single word and matches this word with the dictionary.
This means: unique.string is split wordwise and that word is used for matchning which has lowest osa-distance (all in case of a tie)
example:
"KUESTER and HAUSMEISTER" has distance 0 to both dictString.title HAUSMEISTER and KUESTER. Because the word HAUSMEISTER has minimal distance, another dictString.title HAUMEISTER, which has dist = 1 is not included.
}
\examples{
## Prepare coding index
# write female titles beneath the male title
coding_index <- rbind(coding_index_excerpt[, list(title = bezMale, Code)],
                      coding_index_excerpt[, list(title = bezFemale, Code)])
# standardize titles from the coding index
coding_index <- coding_index[,title := stringPreprocessing(title)]
# drop duplicate lines, might be suboptimal because we keep each title and its associated code only a single time. This means we delete duplicates and the associated, possibly relevant codes.
coding_index <- coding_index[!duplicated(title)]

(x <- c("Abgeordneter", "Abgeordneter", "Abgeordnete", "abgeordnet", "abgeordnet zu xxx", "FSJ", "FSJ2", "Industriemechaniker", "Dipl.-Ing. - Agrarwirtschaft (Landwirtschaft)"))
createSimilarityTableWordwiseStringdist(unique.string = stringPreprocessing(x),
                                        coding_index_w_codes = coding_index,
                                         coding_index_without_codes = frequent_phrases,
                                         preprocessing = list(stopwords = tm::stopwords("de"), stemming = NULL, strPreprocessing = TRUE, removePunct = FALSE),
                                         dist.control = list(method = "osa", weight = c(d = 1, i = 1, s = 1, t = 1)),
                                         threshold = 1)
}
\seealso{
\code{\link{trainSimilarityBasedReasoning}}, \code{\link{createSimilarityTableStringdist}}, \code{\link{createSimilarityTableSubstring}}
}
