% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prepare_German_coding_index_Gesamtberufsliste_der_BA.R
\name{prepare_German_coding_index_Gesamtberufsliste_der_BA}
\alias{prepare_German_coding_index_Gesamtberufsliste_der_BA}
\title{Prepares the Gesamtberufsliste der BA to be used with this package}
\usage{
prepare_German_coding_index_Gesamtberufsliste_der_BA(
  path_to_file,
  count.categories = FALSE
)
}
\arguments{
\item{path_to_file}{path to downloaded file}

\item{count.categories}{(default: \code{FALSE}). The default sets the column \code{count_categories} to 0, avoiding lengthy computations. With \code{count.categories = TRUE} an indicator of job title ambiguity is calculated. See below.}
}
\value{
a data.table with columns Berufsbenennungen, bezMale, and bezFemale, Code, and count_categories.
}
\description{
To use this function go to \url{https://download-portal.arbeitsagentur.de/files/} and download the file \code{Gesamtberufsliste_der_BA.xlsx}. This function prepares a coding index from the Excel file, useful for other functions in this package.
}
\details{
Part of the German Classification of Occupation (KldB 2010) is an alphabetic directory of job titles, available from \url{https://statistik.arbeitsagentur.de/Navigation/Statistik/Grundlagen/Klassifikation-der-Berufe/KldB2010/Systematik-Verzeichnisse/Systematik-Verzeichnisse-Nav.html}. Almost 28.000 job titles are listed in the alphabetic directory along with their associated codes. However, using the alphabetic directory for automatic coding is difficult because it consists of neutral job titles (e.g. Digital-Media-Designer/in) only. They are neutral because the '/' divides the male and female spelling. However, people will never use neutral titles that contain a '/' when they describe their own job. An alternative is the \code{Gesamtberufsliste_der_BA.xlsx}, which contains separate male and female job titles next to the neutral title. The Gesamtberufsliste is used for yearly updates of the alphabetic directory, entailing that both documents are mostly identical.

This functions loads the \code{Gesamtberufsliste_der_BA.xlsx} into memory. It corrects some errors in the Gesamtberufsliste, may remove a few entries, renames some column names, calculates an indicator of job title ambiguity, and, most importantly, it removes parentheses and text in between. The resulting data.table can be used with the function \code{\link{predictWithCodingIndex}} for coding. If coded training data from earlier studies is available, \code{\link{trainSimilarityBasedReasoning}} is another option for using this coding index.

With \code{count.categories = TRUE} an indicator of job title ambiguity is calculated. See the source code for examples what exactly is being calculated. This indicator is only needed if one wishes to change the default behaviour in \code{\link{predictWithCodingIndex}} by setting \code{max.count.categories}. This parameter allows to exclude ambiguous job titles (those with large values in count_categories) from the coding index.
}
\examples{
\dontrun{
# Illustrative results from this function (if you have not downloaded the file yet).
View(coding_index_excerpt)

# point path_to_file to your local file
path_to_file <- ".../Gesamtberufsliste_der_BA.xlsx"
coding_index <- prepare_German_coding_index_Gesamtberufsliste_der_BA(path_to_file, count.categories = FALSE)
}
}
\seealso{
\code{\link{coding_index_excerpt}}, \code{\link{predictWithCodingIndex}}, \code{\link{trainSimilarityBasedReasoning}}
}
