% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringPreprocessing.R
\encoding{UTF-8}
\name{stringPreprocessing}
\alias{stringPreprocessing}
\title{Preprocess German occupational text}
\usage{
stringPreprocessing(verbatim, lang = "de")
}
\arguments{
\item{verbatim}{a character vector.}

\item{lang}{(default de) Everything else will throw an error.}
}
\value{
the same character vector after processing
}
\description{
Function replaces some common characters / character sequences (e.g., Ä, Ü, "DIPL.-ING.") with their uppercase equivalents and removes punctuation, empty spaces and the word "Diplom".
}
\details{
\code{\link{charToRaw}} helps to find UTF-8 characters.
}
\examples{
(x <- c("Verkauf von B\xfcchern, Schreibwaren", "Fach\xe4rzin f\xfcr Kinder- und Jugendmedizin im \xf6ffentlichen Gesundheitswesen", "Industriemechaniker", "Dipl.-Ing. - Agrarwirtschaft (Landwirtschaft)"))
stringPreprocessing(x)
}
