% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainCreecysMemoryBasedReasoning.R
\name{trainCreecysMemoryBasedReasoning}
\alias{trainCreecysMemoryBasedReasoning}
\title{Train Creecys Memory-based reaoning model}
\usage{
trainCreecysMemoryBasedReasoning(data, preprocessing = list(stopwords =
  character(0), stemming = NULL, strPreprocessing = TRUE, removePunct =
  TRUE))
}
\arguments{
\item{data}{a data.table created with \code{\link{removeFaultyAndUncodableAnswers_And_PrepareForAnalysis}}}

\item{preprocessing}{a list with elements
\describe{
  \item{stopwords}{a character vector, use \code{tm::stopwords("de")} for German stopwords.}
  \item{stemming}{\code{NULL} for no stemming and \code{"de"} for stemming using the German porter stemmer.}
  \item{strPreprocessing}{\code{TRUE} if \code{\link{stringPreprocessing}} shall be used.}
  \item{removePunct}{\code{TRUE} if \code{\link[tm]{removePunctuation}} shall be used.}
}}
}
\value{
a processed feature matrix to be used in \code{\link{predictCreecysMemoryBasedReasoning}}
}
\description{
The function does some preprocessing and calculates the importance of various features.
}
\examples{
# set up data
data(occupations)
allowed.codes <- c("71402", "71403", "63302", "83112", "83124", "83131", "83132", "83193", "83194", "-0004", "-0030")
allowed.codes.titles <- c("Office clerks and secretaries (without specialisation)-skilled tasks", "Office clerks and secretaries (without specialisation)-complex tasks", "Gastronomy occupations (without specialisation)-skilled tasks",
 "Occupations in child care and child-rearing-skilled tasks", "Occupations in social work and social pedagogics-highly complex tasks", "Pedagogic specialists in social care work and special needs education-unskilled/semiskilled tasks", "Pedagogic specialists in social care work and special needs education-skilled tasks", "Supervisors in education and social work, and of pedagogic specialists in social care work", "Managers in education and social work, and of pedagogic specialists in social care work",
 "Not precise enough for coding", "Student assistants")
proc.occupations <- removeFaultyAndUncodableAnswers_And_PrepareForAnalysis(occupations, colNames = c("orig_answer", "orig_code"), allowed.codes, allowed.codes.titles)

# Recommended configuration (and commonly used in this package)
memModel <- trainCreecysMemoryBasedReasoning(proc.occupations,
                 preprocessing = list(stopwords = character(0), stemming = NULL, strPreprocessing = TRUE, removePunct = FALSE))
}
\seealso{
\code{\link{predictCreecysMemoryBasedReasoning}}

Creecy, R. H., Masand, B. M., Smith, S. J., Waltz, D. L. (1992). Trading MIPS and Memory for Knowledge Engineering. Comm. ACM 35(8). pp. 48--65.
}
